app/requirements.txt
alembic==1.11.3 ; python_version >= "3.10" and python_version < "4.0"
apispec[yaml]==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
click==7.1.2 ; python_version >= "3.10" and python_version < "4.0"
combojsonapi==1.1.2 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.10" and python_version < "4.0"
flask-admin==1.6.1 ; python_version >= "3.10" and python_version < "4.0"
flask-bcrypt==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
flask-combo-jsonapi==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
flask-login==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
flask-migrate==4.0.4 ; python_version >= "3.10" and python_version < "4.0"
flask-sqlalchemy==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
flask-wtf==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
flask==1.1.4 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==1.1.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==2.11.3 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
marshmallow-jsonapi==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.7 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
simplejson==3.19.1 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.3.24 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
wtforms==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
